[IMPROVE] TokenRepository 추상화 및 적용 #43
